Adding new Site
Added new site projects to test if the entry points in vite.config.js
are correct. Because of build.yml I chose build as the outDir since
that is the name of the branch as well. I moved the index.html into
the src and the icon into the src/assets.